Moving more classes out of framework
