+++ src/mantle/new-relic/composer.json
@@ -1,0 +1,33 @@
+{
+    "name": "mantle-framework/new-relic",
+    "description": "The Mantle Framework New Relic Package",
+    "type": "project",
+    "require": {
+        "php": "^8.0",
+        "alleyinteractive/composer-wordpress-autoloader": "^1.0",
+        "mantle-framework/contracts": "^0.10",
+        "mantle-framework/http": "^0.10",
+        "mantle-framework/support": "^0.10"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "0.10-dev"
+        },
+        "wordpress-autoloader": {
+            "autoload": {
+                "Mantle\\New_Relic": "./"
+            }
+        }
+    },
+    "license": "GPL-2.0-or-later",
+    "authors": [
+        {
+            "name": "Alley",
+            "email": "info@alley.co"
+        }
+    ],
+    "minimum-stability": "dev",
+    "config": {
+        "sort-packages": true
+    }
+}
